# sdks/sdkwork-birdcoder2-app-sdk/openapi/sdkwork-birdcoder2-app-api.openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "SDKWork BirdCoder2 App API",
    "version": "0.1.0",
    "description": "Canonical BirdCoder2 multi-host registry and relayed agent conversation App API authority. The mobile H5 client registers hosts, issues pairing codes, and drives agent conversations that run on a Windows, Linux, macOS, Docker, or cloud sdkwork-sandbox host."
  },
  "servers": [
    {
      "url": "/app/v3/api"
    }
  ],
...
  "paths": {
    "/app/v3/api/birdcoder/hosts": {
      "get": {
        "operationId": "hosts.list",
        "summary": "List the caller's hosts",
        "description": "List the hosts the authenticated user owns, newest first. Cursor paginated; the cursor is opaque and stable across pages.",
        "tags": [
          "hosts"
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque cursor returned as nextCursor by the previous page.",
            "schema": {
              "type": "string",
              "maxLength": 512
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "description": "Requested page size; default 20, maximum 100.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Host page returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostListEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Problem response envelope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-data-scope": "owner",
        "x-sdkwork-deployment": "all",
        "x-sdkwork-domain": "hosts",
        "x-sdkwork-permission": "birdcoder.hosts.read",
        "x-sdkwork-public": false,
        "x-sdkwork-resource": "birdcoder.host",
        "x-sdkwork-surface": "app",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-api-authority": "sdkwork-birdcoder2-app-api",
        "x-sdkwork-owner": "sdkwork-birdcoder2",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      }
    },
...
    "/app/v3/api/birdcoder/hosts/{hostId}": {
      "get": {
        "operationId": "hosts.retrieve",
        "summary": "Get one host",
        "description": "Returns one host the authenticated user owns, including its live lease state.",
        "tags": [
          "hosts"
        ],
        "parameters": [
          {
            "name": "hostId",
            "in": "path",
            "required": true,
            "description": "Host identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Host returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Problem response envelope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-data-scope": "owner",
        "x-sdkwork-deployment": "all",
        "x-sdkwork-domain": "hosts",
        "x-sdkwork-permission": "birdcoder.hosts.read",
        "x-sdkwork-public": false,
        "x-sdkwork-resource": "birdcoder.host",
        "x-sdkwork-surface": "app",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-api-authority": "sdkwork-birdcoder2-app-api",
        "x-sdkwork-owner": "sdkwork-birdcoder2",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      },
      "patch": {
        "operationId": "hosts.update",
        "summary": "Rename, relabel, or administratively disable one host",
        "description": "Partial update. Only the fields present in the body change. status accepts offline or disabled; online and pending are lifecycle-owned and rejected.",
        "tags": [
          "hosts"
        ],
        "parameters": [
          {
            "name": "hostId",
            "in": "path",
            "required": true,
            "description": "Host identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Host updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Problem response envelope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-data-scope": "owner",
        "x-sdkwork-deployment": "all",
        "x-sdkwork-domain": "hosts",
        "x-sdkwork-permission": "birdcoder.hosts.write",
        "x-sdkwork-public": false,
        "x-sdkwork-resource": "birdcoder.host",
        "x-sdkwork-surface": "app",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-api-authority": "sdkwork-birdcoder2-app-api",
        "x-sdkwork-owner": "sdkwork-birdcoder2",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      },
      "delete": {
        "operationId": "hosts.delete",
        "summary": "Remove one host",
        "description": "Removes the host and every conversation it carried. The host runtime loses its lease immediately.",
        "tags": [
          "hosts"
        ],
        "parameters": [
          {
            "name": "hostId",
            "in": "path",
            "required": true,
            "description": "Host identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Host removed."
          },
          "default": {
            "description": "Problem response envelope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-data-scope": "owner",
        "x-sdkwork-deployment": "all",
        "x-sdkwork-domain": "hosts",
        "x-sdkwork-permission": "birdcoder.hosts.write",
        "x-sdkwork-public": false,
        "x-sdkwork-resource": "birdcoder.host",
        "x-sdkwork-surface": "app",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-api-authority": "sdkwork-birdcoder2-app-api",
        "x-sdkwork-owner": "sdkwork-birdcoder2",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token",
        "description": "Signed JWT access_token compact serialization (header.payload.signature)."
      }
    },
    "schemas": {
      "SdkWorkApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "description": "Server-owned request correlation id."
          }
        },
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "additionalProperties": false
      },
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "minimum": 40001
          },
          "traceId": {
            "type": "string",
            "description": "Server-owned request correlation id."
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "additionalProperties": true
      },
      "PageInfo": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "offset",
              "cursor"
            ]
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "mode",
          "pageSize",
          "hasMore"
        ],
        "additionalProperties": false
      },
      "HostPlatform": {
        "type": "string",
        "description": "Machine kind the host runs on.",
        "enum": [
          "windows",
          "linux",
          "macos",
          "docker",
          "cloud-sandbox"
        ]
      },
      "HostStatus": {
        "type": "string",
        "description": "Reachability of a host. pending: registered, no runtime attached. online: a runtime holds a live lease. offline: the lease lapsed or was closed. disabled: an operator disabled it.",
        "enum": [
          "pending",
          "online",
          "offline",
          "disabled"
        ]
      },
      "HostAdministrativeStatus": {
        "type": "string",
        "description": "The subset of host status an owner may set directly.",
        "enum": [
          "offline",
          "disabled"
        ]
      },
...
      "EpochSeconds": {
        "type": "string",
        "format": "int64",
        "pattern": "^[0-9]+$",
        "description": "Instant in Unix epoch seconds. int64 crosses the wire as a string (API_SPEC section 13.6).",
        "x-sdkwork-int64-string": true
      },
...
      "Host": {
        "type": "object",
        "description": "One machine or runtime running a sdkwork-birdcoder2 instance.",
        "properties": {
          "hostId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "platform": {
            "$ref": "#/components/schemas/HostPlatform"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "runtimeVersion": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/HostStatus"
          },
          "leaseId": {
            "type": [
              "string",
              "null"
            ]
          },
          "leaseExpiresAt": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/EpochSeconds"
              },
              {
                "type": "null"
              }
            ]
          },
          "lastSeenAt": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/EpochSeconds"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "$ref": "#/components/schemas/EpochSeconds"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/EpochSeconds"
          }
        },
        "required": [
          "hostId",
          "displayName",
          "platform",
          "labels",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
      "HostUpdateRequest": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 120
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "$ref": "#/components/schemas/HostAdministrativeStatus"
          }
        },
        "additionalProperties": false
      },
...
      "HostEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/Host"
                  }
                },
                "required": [
                  "item"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "data"
            ],
            "additionalProperties": false
          }
        ]
      },
      "HostListEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Host"
                    }
                  },
                  "pageInfo": {
                    "$ref": "#/components/schemas/PageInfo"
                  }
                },
                "required": [
                  "items",
                  "pageInfo"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "data"
            ],
            "additionalProperties": false
          }
        ]
      },
...
    }
  }
}